REF/TST: implement test_interpolate for Series (#32112)
